+++ src/main/scala/factorio_data/Recipe.scala
@@ -5,6 +5,6 @@ class Recipe(val id: String,
              val time: Double,
              val producers: Array[Machine],
              val in: Map[Item, Int],
-             val out: Map[Item, Int],
+             val out: Map[Item, Double],
              disallowedEffects: Array[String],
              catalyst: Map[Item, Int])
